created npm package and made styling edits
